src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/flat-tree-traversal-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Tests for Composed Shadow DOM Tree Traversal APIs. Can only run within DRT

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

ShadowRoot should be used.
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
